fix: replaced uint8 with uint256

+++ src/Facets/StargateFacet.sol
@@ -64,8 +64,8 @@ contract StargateFacet is ILiFi, SwapperV2, ReentrancyGuard {
         _stargateData.amountLD = _executeAndCheckSwaps(_lifiData, _swapData, payable(msg.sender));
 
         uint256 nativeFee = msg.value;
-        uint8 nSwaps = uint8(_swapData.length);
-        for (uint8 i = 0; i < nSwaps; ) {
+        uint256 nSwaps = _swapData.length;
+        for (uint256 i = 0; i < nSwaps; ) {
             if (LibAsset.isNativeAsset(_swapData[i].sendingAssetId)) {
                 nativeFee -= _swapData[i].fromAmount;
             }
